# lib/api-zod/src/generated/types/album.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Jannat Events public website and content management API
 * OpenAPI spec version: 0.1.0
 */

export interface Album {
  id: number;
  title: string;
  slug: string;
  description: string;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  eventId?: number | null;
  /** @nullable */
  eventTitle?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  date?: Date | null;
  published: boolean;
  createdAt: Date;
}