# lib/api-zod/src/generated/types/homePageUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Jannat Events public website and content management API
 * OpenAPI spec version: 0.1.0
 */

export interface HomePageUpdate {
  /** @nullable */
  heroImage?: string | null;
  /** @nullable */
  heroVideo?: string | null;
  heroHeadline?: string;
  heroSubtitle?: string;
  ctaText?: string;
  ctaUrl?: string;
  /** @nullable */
  featuredEventId?: number | null;
}